refactor: move fs helper fns, isLocked -> isProfileLocked
