+++ packages/zowe-explorer-api/src/utils/FileManagement.ts
@@ -13,6 +13,8 @@ import { realpathSync } from "fs";
 import { platform } from "os";
 import { Constants } from "../globals";
 import { ImperativeConfig, ConfigUtils } from "@zowe/imperative";
+import { IFileSystemEntry, ZoweScheme } from "../fs/types";
+import { window, workspace } from "vscode";
 
 export class FileManagement {
     public static permStringToOctal(perms: string): number {
@@ -49,5 +51,35 @@ export class FileManagement {
             }
         }
         return realpathSync(anyPath);
+    }
+
+    public static async reloadActiveEditorForProfile(profileName: string): Promise<void> {
+        if (
+            (Object.values(ZoweScheme) as string[]).includes(window.activeTextEditor.document.uri.scheme) &&
+            window.activeTextEditor.document.uri.path.startsWith(`/${profileName}/`) &&
+            !window.activeTextEditor.document.isDirty
+        ) {
+            const fsEntry = (await workspace.fs.stat(window.activeTextEditor.document.uri)) as IFileSystemEntry;
+            fsEntry.wasAccessed = false;
+            await workspace.fs.readFile(window.activeTextEditor.document.uri);
+        }
+    }
+
+    public static async reloadWorkspacesForProfile(profileName: string): Promise<void> {
+        const foldersWithProfile = (workspace.workspaceFolders ?? []).filter(
+            (f) => (f.uri.scheme === ZoweScheme.DS || f.uri.scheme === ZoweScheme.USS) && f.uri.path.startsWith(`/${profileName}/`)
+        );
+        for (const folder of foldersWithProfile) {
+            try {
+                await workspace.fs.stat(folder.uri.with({ query: "fetch=true" }));
+            } catch (err) {
+                if (err instanceof Error) {
+                    // TODO: Remove console.error in favor of logger
+                    // (need to move logger to ZE API)
+                    // eslint-disable-next-line no-console
+                    console.error("reloadWorkspacesForProfile:", err.message);
+                }
+            }
+        }
     }
 }
